Replace old OpenAPI preview tool with new one
