+++ src/lib/get-breadcrumb-links.ts
@@ -37,7 +37,6 @@ const KNOWN_URL_TITLE: Record<string, string> = {
 const KNOWN_URL_SEGMENT_TITLE: Record<string, string> = {
 	'api-docs': 'API',
 	'vault-secrets': 'Vault Secrets',
-	'open-api-docs-preview': 'OpenAPI Docs Preview Tool',
 }
 
 /**
