--- src/pages/open-api-docs-preview-v2/[[...page]].tsx
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import OpenApiDocsPreviewViewV2 from 'views/open-api-docs-preview-v2'
-import { getServerSideProps } from 'views/open-api-docs-preview-v2/server'
-
-export { getServerSideProps }
-export default OpenApiDocsPreviewViewV2
