+++ src/pages/open-api-docs-preview/[[...page]].tsx
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+import OpenApiDocsPreviewView from 'views/open-api-docs-preview'
+import { getServerSideProps } from 'views/open-api-docs-preview/server'
+
+export { getServerSideProps }
+export default OpenApiDocsPreviewView
