--- src/views/open-api-docs-preview-v2/components/file-string-input/file-string-input.module.css
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-.root {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-}
-
-.helperText {
-	color: var(--token-form-helper-text-color);
-}
