--- src/views/open-api-docs-preview-v2/components/textarea-input/textarea-input.module.css
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-.root {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-}
-
-.helperText {
-	color: var(--token-form-helper-text-color);
-}
-
-.textarea {
-	width: 100%;
-	resize: vertical;
-	min-height: 200px;
-}
